+++ deep_tempering/__init__.py
@@ -1,3 +1,4 @@
 from deep_tempering.training import EnsembleModel
 from deep_tempering.training_test import model_builder as example_model_builder
 from deep_tempering import *
+from deep_tempering.training_utils import load_optimal_model
