+++ deep_tempering/training_test.py
@@ -6,6 +6,7 @@ import numpy as np
 from sklearn.datasets import make_blobs
 
 from deep_tempering import training
+from deep_tempering import training_utils
 from deep_tempering import callbacks as cbks
 
 def test_hp_space_state():
@@ -20,7 +21,7 @@ def test_hp_space_state():
           'learning_rate': np.linspace(0.001, 0.01, n_replicas),
           'dropout_rate': np.linspace(0., 0.6, n_replicas)
       }
-  hpss = training.HyperParamSpace(em, hparams_dict)
+  hpss = training_utils.HyperParamSpace(em, hparams_dict)
 
   # test that initial hyper-parameter values are correct
   initial_values = {
